Add offset transformation for pressure sensors (see desc)
Not sure why, but the analog reading value is consistently slightly less than it should be. Originally I though something was wrong with the voltage divider or software, but it appears to be an issue internal to the ESP32/ADC. When I measured 0.50V from the sensor on the analog pin, the ESP32 would report an analog reading that corresponded to an input voltage of ~0.32. This result carried through for 1V, 2V, 3V, etc. It appears to be linearly incorrect, so I added an offset function that ensures the pin voltage is accurately reconstructed after ADC processing